rename queue.h to mqtt_queue.h to avoid confict with sdk
NONOS SDK 2.1.0 already contains another queue.h which is
conflicting.
