+++ mqtt/queue.c
@@ -27,7 +27,7 @@
 * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 * POSSIBILITY OF SUCH DAMAGE.
 */
-#include "queue.h"
+#include "mqtt_queue.h"
 
 #include "user_interface.h"
 #include "osapi.h"
